added imgur search into helpers folder and moved commands to cogs
